+++ changelog.md
@@ -35,6 +35,7 @@ After:
 - ✨ Improved feedback loop node, you can now apply effects after the feedback node, that won't affect the image sent back into the loop
 - 🐛  When a camera node was present but not actually linked into the graph, it still used the camera and forced us to rerender every frame. Now only the nodes that are actually used affect us.
 - Added server, can send http request to set values in Coollab
+- 🤏 Merged Camera and Inspector windows. You can now edit camera values in the Inspector, when no node is selected.
 
 ## 🐣beta-18